third_party/WebKit/LayoutTests/media/video-controls-show-on-focus.html

<!DOCTYPE html>
<html>
<title>Test visibiblity of controls when focusing of &lt;video></title>
<script src=video-test.js></script>
<script src=media-controls.js></script>
<p>
Test that video controls re-appear when the media (&lt;video>) element is focused.
</p>
<video controls autoplay src="content/test.ogv"></video>
<script>
var controls;
var video = document.querySelector("video");

video.addEventListener("playing", function()
{
    runAfterControlsHidden(function()
    {
        controls = mediaControlsButton(video, "panel");

        testExpected("getComputedStyle(controls).opacity", 0);

        video.focus();

        setTimeout(function()
        {
            testExpected("getComputedStyle(controls).opacity", 1);

            consoleWrite("");
            endTest();
        }, controlsFadeInDurationMs);
    }, video);
});
</script>
